change var type

+++ src/main/java/lang/print/gaps/finalModuleTask/TemperatureConverter.java
@@ -2,8 +2,8 @@ package lang.print.gaps.finalModuleTask;
 
 public class TemperatureConverter {
     public void toFahrenheit(int temperatureCelsius){
-        int fahrenheit;
-        fahrenheit = (temperatureCelsius * (int)1.8) + 32;
+        double fahrenheit;
+        fahrenheit = (temperatureCelsius * 1.8) + 32;
         System.out.println(fahrenheit);
     }
 }
